Add TransactionDetailFragment

+++ app/src/main/java/com/stephengrice/laser/db/DbHelper.java
@@ -60,7 +60,7 @@ public class DbHelper extends SQLiteOpenHelper {
         while (cursor.moveToNext()) {
             DbContract.Transaction current = new DbContract.Transaction(
                     cursor.getFloat(cursor.getColumnIndexOrThrow(DbContract.Transaction.COLUMN_NAME_AMOUNT)),
-                    cursor.getInt(cursor.getColumnIndexOrThrow(DbContract.Transaction.COLUMN_NAME_DATE)),
+                    cursor.getLong(cursor.getColumnIndexOrThrow(DbContract.Transaction.COLUMN_NAME_DATE)),
                     cursor.getString(cursor.getColumnIndexOrThrow(DbContract.Transaction.COLUMN_NAME_DESCRIPTION)),
                     cursor.getInt(cursor.getColumnIndexOrThrow(DbContract.Transaction.COLUMN_NAME_CATEGORY_ID)),
                     cursor.getString(cursor.getColumnIndexOrThrow(DbContract.Category.COLUMN_NAME_TITLE))
